Fix gcc error: no matching function for call to 'expect_buffer(aeon::streams::aggregate_device<aeon::streams::sink_buffer_filter<10>, aeon::streams::memory_device<std::vector<char> > >&, const std::array<char, 10>&)'

+++ src/streams/tests/test_libaeon_streams/test_buffer_filter.cpp
@@ -10,7 +10,7 @@
 
 using namespace aeon;
 
-template <typename pipeline_t, int size>
+template <typename pipeline_t, long unsigned int size>
 void expect_buffer(pipeline_t &pipeline, const std::array<char, size> &expected)
 {
     std::array<char, size> data_readback{};
